Add solutions to 02-class-id-selectors
Add odd class to odd number of element,
and made it lightpink background color in rgb value.
and font family of verdana, dejavu sans and sans serif.
Add odd-cool class to 3rd element,
and made it 24px font.
Add id of two to 2nd element,
and made the text blue with 36px font.
Add id of four to 4th element,
and made it light green in hsl value,
and 24px and bold font.